Final Year Project
In this project we used BERT (Bidirectional Encoder Representations from Transformers) which a natural language processing framework introduced by google in 2018. The dataset that we used in this project is Balochi Named Entity Recognition Dataset which is tagged in INCEpTION.
INCEpTION is an open source jave based text annotation flatform.